IC-281-Refactor moved ResultSet and ResultRow to db
