+++ utils/db.ts
@@ -1,6 +1,7 @@
 import { pipe } from "fp-ts/lib/function";
 import { DatabaseError, Pool, QueryResult } from "pg";
 import * as TE from "fp-ts/TaskEither";
+import * as t from "io-ts";
 import { IDecodableConfigPostgreSQL } from "./config";
 
 // eslint-disable-next-line functional/no-let
@@ -31,5 +32,17 @@ export const queryDataTable = (
       error => error as DatabaseError
     )
   );
+
+export const ResultRow = t.interface({
+  count: t.string,
+  status: t.string
+});
+export type ResultRow = t.TypeOf<typeof ResultRow>;
+
+export const ResultSet = t.interface({
+  rowCount: t.number,
+  rows: t.readonlyArray(ResultRow)
+});
+export type ResultSet = t.TypeOf<typeof ResultSet>;
 
 export default getPool;
